Do not use GtkTargetPair and remove it.
It was used in DropTarget.setEventData where the list was iterated using
g_list_nth and GtkTargetPair struct recreated on the java side. It is
way more simple and effective to use g_list_nth_data which gives us
exactly what we need, reduces java/c roundtrips and allows to be
compiled on GTK 3.