Fix for initial bounds not set on GTK 3.
There is the following printed on console:
Gtk-CRITICAL **: gtk_widget_set_allocation: assertion
`gtk_widget_get_visible (widget) || gtk_widget_is_toplevel (widget)'
failed. The critical part is that this breaks even the workspace
selection dialog (it's empty) on GTK 3.

A workaround is to make the widget visible. It's done for GTK 3 only to
prevent any possible problem on GTK 2 because of it.