Bug 417601 - [GTK3] Menu items are missing keyboard shortcuts
It is a matter of connecting the GtkAccelLabel to the handle in order to
get the shortcuts in the way GTK represent them. This doesn't allow
custom formatting of the shortcut text but GTK does not provide an API
to achieve that. What worked for GTK 2.x was simply abusing the fact
that structs were all public not considering sealing at all. 
Bonus point for shortcut text being consistent with other GTK
applications.

Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>