Bug 441705 - Check cairo version to enable USE_CAIRO
org.eclipse.swt.widgets.Caret uses CAIRO_OPERATOR_DIFFERENCE when
OS.USE_CAIRO is true but that operator is not introduced until
cairo 1.9.4. OS.USE_CAIRO is set based on the GTK version which
leads the boolean to be enabled on systems where GTK >= 2.24 and
if cairo is < 1.9.4 eclipse crashes. This patch checks the cairo
version when enabling cairo.

Change-Id: I183adafba7ed95da0c6ec1cf05b8a52b7428b1ca
Signed-off-by: Sami Wagiaalla <swagiaal@redhat.com>