Bug 444303 - UI Responsiveness Monitoring produces false positives on Mac Cocoa
Replaced SWT.Sleep and SWT.Wakeup events by more general
SWT.PreExternalEventDispatch and SWT.PostExternalEventDispatch events
and added sending of these events in places that previously required
filters. With this approach the remaining filter for Cocoa can be
removed from MonitoringPreferenceInitializer. The filters on other
platforms can also be removed once
sendPreExternalEventDispatchEvent/sendPostExternalEventDispatchEvent
calls are added to the methods that are currently listed in the filters
or to the methods calling them.

Change-Id: Iaca1b5e4021be40f3e9fcbb079cf99d1b6dba069
Signed-off-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>