Bug 449546 - [responsiveness] No error logged when UI froze during opening type hierarchy on java.lang.Object on Mac
My previous approach to event filtering turned out to be unreliable
since the original event type is not guaranteed to be preserved after
the event is processed. The new approach is to make the type of the
original event available in the "detail" field of the pre and post
events. 

Change-Id: Ia592a754d97bb4654ad71d88a9ad48714e45be7c
Signed-off-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>