Bug 457041 - [GTK] Trailing whitespace removal for GTK codebase
I traversed SWT's GTK-related code to remove trailing whitespace.

I ran the junit tests (gkt2/gtk3) to ensure everything still works as
before. (that said there are 3 Junit test erros if ran with GTK3, but
that is not related to this whitespace fix).

ControlExample and other common test widgets work as usual.

Change-Id: I49aaeba65ed693c26c580d31014f78ec57068ea8
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>