Bug 445456 - Ignored style hints for shells
Shells with a parent can't have any button on the titlebar except for
the close one, when running Metacity. This fix enables the minimize and
maximize buttons for child shells, by removing the
GDK_WINDOW_TYPE_HINT_DIALOG type hint and calling the
gdk_window_set_functions method when setting the decorations of a shell.

The code was tested against CentOS release 6.6, Metacity 2.28.0 and
gtk2-2.2423-6.

Change-Id: If320be2ac9ecd7fe3090a164add1015b4ff0f228
Signed-off-by: Alexandra Buzila <abuzila@eclipsesource.com>