Bug 461616 -  Combo improvements in background/foreground for gtk3
Incremental change to make Combo use rgba for foreground on Gtk3.
This also fixes combo's background for Gtk3, so that the Junit test now
works in gtk3.

object_set_..rgba required special attention. I added a line to os.h so
that it would also compile on gtk2 (as the GdkRGBA in the paramater list
is gtk3 only)


Change-Id: Iacb608f773f54028bfacbb7d1e0b15d0567d14f1
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>