Bug 462009- [Gtk3] Button background Color fix
In Gtk3, the background of a button does not get set correctly when button.setBackground() is called, theaming doesn't work properly. 

This patch fixes the 'setting' part of it. Please see screen shots of the related bug.

There is still an outstanding issue with *retrieving* the correct background color, but that will be addressed in a subsequent patch.

Change-Id: I4970da1e7ad3c19916116dc0095d5088917bbf0c
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>