Bug 461427 - SWT fragment missing on classpath in build of SWT tests andrefs/changes/49/43849/1 tools when moving to tycho 0.23.0-SNAPSHOT
Property org.eclipse.swt.buildtime=true needs to be define only in
build-individual-bundles as the repo is not buildable if this profile is
not enabled anyway.
This shouldn't have been reverted as it is not affecting
releng.aggregator build and losens the buildtime requirements to at
least allow resolving with later failing compilation.

Change-Id: I7ca5c9a09eae64b02a5cc5cf2d6d5412fcf9c0c2
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>