Bug 464816 - [Gtk3] Adding SWT.EraseItem listener and drawing background makes the TreeItem/TableItem text dissapear - Updated Code for Table as well
The code change is related to setting the correct clipping during the EraseItem event.

This has been tested on Fedora 19 with GTK+ 3.8.8 and Ubuntu 14.04 with GTK+ 3.10.

Testing is done using the snippet provided and modified snippet for table as well. Ran the SWT junit test suite. 

For regression testing, used a self hosted eclipse and tested on multiple views like Call hierarchy, Type hierarchy, etc. Verified that they are working fine.

Change-Id: I8bdb7f219ccbfec467b85c88a1e72c5fdc3dcf3e
Signed-off-by: Niraj Modi <niraj.modi@in.ibm.com>
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>