Bug 467687 - SWT: Invalid thread access in GC.drawImage
On Windows created a new method getActualDPI() that works in non ui
threads as well and using this to get the dpi

On linux modified the getActualdpi() method to not use getbounds call
which is actually not required as we need on only x and y of the
display. these are always 0, 0 in all cases of getBounds.

Updated with review comments

Change-Id: I7154ca7a2493fc21d1e38388a2f19c885ce9e4bf
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>