Bug 469521: Replace deprecated gtk_container_resize_children()
Tested on both Gtk3 (3.16) and Gtk2 (2.24). Used TableTest.java
from bug 226973 to test this proposed patch. Now when the table
is not visible, all column widths return the correct value when
queried using getWidth() (as the previous fix had implemented).

Table.java was the only file that had a call to 
gtk_container_resize_children. The solution was to replace
this call with a call to the forceResize() method.

This patch is ready for review.

Change-Id: I0a5e30dff2d5cb28a9ea03a57ab99ed81bf89f5f
Signed-off-by: Eric Williams <ericwill@redhat.com>