Bug 312568 (Part 1) - [GTK] Editor does not get focus after double click from Project/Package Explorer
This is part 1 of the fix to prevent package explorer from stealing
focus.

Here I remove a workaround introduced 11 years ago. See:
git show 300ace8e3eaec6136b4861feec65efae8fe149fc

Originally not-returning focus caused a seg-fault in GtkTree.
But this is no longer the case with current Gtk.

TESTS PERFORMED after removal:
- 64 bit machine
- (SWT Tree+List+Table) * (Gtk2.24 + Gtk3.14)
-> No crashes observed.

I Used snippet in my Repo for testing:
https://github.com/LeoUfimtsev/LeoSwtTests/blob/master/src/standaloneTests/LTEST_465984_TreeDoubleClick.java
But any Tree/Table/List snippet will do.

Noticible changes in the behaviour of the snippet:
 - GTK2 : No changes can be observed with/without patch.
 - GKT3 : Focus now stays in the text field after double-clicking on
          SWT Tree/List/Table.

Part 2 of the fix is on it's way.

Change-Id: Ie27503caac60a97459b3cb86dc9e024f2e879cd0
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>