Bug 470298: [Gtk3] Replace deprecated gtk_image_menu_item_new_with_label
This patch includes latest fixes for both bugs, including updated
Javadocs and fixes to the setImage() method.

This patch ports functionality to gtk_menu_item_new() from the now
deprecated gtk_image_menu_item_new(). This is done by packing a GtkBox
with an image + label (for example), and then adding that box as the
child of the GtkMenuItem container. More information about this solution
can be found in the Gtk3 reference page for GtkImageMenuItem.

Unfortunately, this solution means that this patch is rather large, due
to the number of corner cases to consider since we are using a GtkBox.
Additionally, some exact behavior in Gtk3 could not be replicated
identically to that of Gtk2, so some additional code had to be added --
including an environment variable to support double padding of
MenuItems. Using the environment variable SWT_PADDED_MENU_ITEMS with a
value of 1 will introduce the double padding. By default, MenuItems will
not be double padded. More information about double padded MenuItems can
be found under the comments for this bug:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=470298

Some additional functionality was also added to Gtk3 by virtue of the
fact that we are no longer bound to GtkImageMenuItem. For example, check
and radio buttons now have support for images alongside radios and check
marks (where they previously didn't). In fact, any MenuItem (except for
separators) now supports images/icons along side text/buttons/etc. This
functionality can be accessed using the setImage() method of that
particular MenuItem object.

All previous functionality from Gtk2 has been successfully replicated in
Gtk3. AllNonBrowser JUnit tests pass on Gtk3.16 and Gtk2.24. This patch
was tested using ControlExample.java, TextEditor.java, and a child
Eclipse on the following versions of Gtk: 3.16, 3.14, 3.12, 3.10, 3.8,
and 2.24. This patch was also fully tested on RHEL 7.1, on both Gtk2.24
and Gtk3.8.

Change-Id: Ica575e9cfd0db5775821f67dd1000aab8ee4ce3b
Signed-off-by: Eric Williams <ericwill@redhat.com>