Bug 470308: [Gtk3] ColorDialog.java port and cleanup
Tested on Gtk2 and Gtk3, using Snippet133. Behaviour is as expected.
AllNonBrowser unit tests pass in Gtk2 and Gtk3.

This patch is longer than usual because the method it is contained in
uses both GdkColor and GdkRGBA. While the color selection functionality
was implemented in both Gtk2 and 3, there was no equivalent
functionality for gtk_settings_set_string_property() in Gtk3. Because of
this, I had to rework the method and add functionality for Gtk3 to
mirror that of Gtk2.

Change-Id: I9467255a6ced41e4b36715a88f1de7b6f5ed4531
Signed-off-by: Eric Williams <ericwill@redhat.com>