Bug 471831: [Gtk3] Replace deprecated property GtkMisc:ypad
Patch to replace the deprecated GtkMisc:ypad property.

Tested using Snippet109 by calling computeSize() on the two labels
created in each pane. Calling computeSize() in both Gtk3 and Gtk2
returns the same results. AllNonBrowser JUnit tests pass in Gtk 3 and 
Gtk2.

There was a call in the computeSize() method of Label.java that
retrieved the integer value of the vertical padding (top and bottom) of
the label. In Gtk3, this call will not run and instead will be replaced
with calls to retrieve the replacement "margin-top" and "margin-bottom"
properties.

Change-Id: I448a0a371ebf4e6ba4f43ce1eecf96cb1e523c39
Signed-off-by: Eric Williams <ericwill@redhat.com>