Bug 393409: Tooltip background black on black
Before this patch, tooltips in Gtk3 were not displaying correctly. No
text or background was being drawn, resulting in tooltips being entirely
black. This was due to two reasons:

1) The draw event was not being handled properly in Gtk3.9+. This means
that a draw signal was being emitted, but the signal was not handled at
all, causing the drawTooltip() function in ToolTip.java not to be
called. Therefore, nothing was being rendered/painted, leaving a black
window instead. This was due to a change in draw/expose events in native
Gtk from version 3.9 and up.

2) Once the draw signal was properly connected, only text was being
drawn. The background color of the widget was not properly set, so it
was showing a black background. This was due to a call to
gtk_widget_set_app_paintable() in createHandle() which, when set to
true, forces the widget to paint its background manually instead of
getting it from a style provider or from a parent widget.
gtk_widget_set_app_paintable() was overriding a call to
gtk_widget_override_background_color(), meaning no background color was
actually being set, resulting in a black background.

This patch implements the following behaviour in Gtk3:

Gtk3.10+: the system background color will be applied using CSS styling,
if available. If no system background color is specified, the tooltip
will inherit its background from its parent container/widget. This was
the original intended behaviour.

Gtk3.0-3.9: the system background color will be applied using
gtk_widget_override_color(). Same behaviour of Gtk3.10+, just by virtue
of a GtkWidget call instead of a CSS style provider.

Tested on Gtk2.24, and Gtk3.4+. Used a modified version of Snippet225 to
test a tray "balloon" tooltip. ControlExample was also used in testing.
AllNonBrowser JUnit tests pass in Gtk3 and Gtk2.

Change-Id: I6604d3b6d6773d550639028a3fbc04724fc6276f
Signed-off-by: Eric Williams <ericwill@redhat.com>