Bug 476562: [Gtk3] Wide icons cut off in Tree when using custom drawing (SWT.PaintItem) listener
This change was formerly filed under Bug 469277, but has now been moved
to Bug 476562.

This patch addresses the bug where in certain cases, images in TreeItems
would be cut off. This happened when images were wider than 16px,
resulting in text rendering over the right side of image. This was seen
mainly in the Outline view when enabling styled label providers (Window
-> Preferences -> General -> Appearance -> Use Mixed Fonts for Labels
and Icons). 

The getTextBounds() method in TreeItem.java was causing the issue, as
the geometric area for text renderers was not taking the image size into
consideration, causing the width to be set at 16px. This resulted in
text rendering occurring after 16px, when the image could be wider than
16px. There is now a check in place that re-adjusts the bounds of the
text renderer if the width of the image is larger than the default
offset. 

Tested on Gtk3.16, 3.14, 3.12, 3.10, 3.8, and 2.24. AllNonBrowswer JUnit
tests pass on Gtk3 and Gtk2.

Change-Id: I9d597f859490bbc9c9a3e9262624afde8da7962d
Signed-off-by: Eric Williams <ericwill@redhat.com>