Bug 469277 - [GTK3] Tree cuts wide icons on the right
This patch fixes two bugs: bug 468277 and 476419. 

Icons were being cut off when using native painting because the renderer
for the icon pixbuf was hardcoded at 16px. This also caused gaps between
the expander triangle and text when using Trees with text only. This
patch correctly updates the size of the pixbuf renderer to the size of
the image, and also prevents text only trees from having gaps between
the expander triangle and text. 

Tested on Gtk3.16, Gtk3.14, Gtk3.10, Gtk3.8, and Gtk2.24. AllNonBrowser
JUnit tests pass in both Gtk2 and Gtk3.

Change-Id: Ib820827c2e1accdf046b0e0079b7bf2625dd83fc
Signed-off-by: Eric Williams <ericwill@redhat.com>