Bug 478560: [Graphics] [Import/Export] Wizard has some entries with large icon sizes
Combined dual bug fix: Bug 478915 depends on fixes for Bug 478650 (this
bug), so the changes for both have been merged into a single patch.

Fixes for Bug 478560: this deals with how we size pixbufRenderers for
images in TreeItems. To maintain consistency with other platforms, we
now set the size of pixbufRenderers to the size of the first image set.
A boolean, pixbufSizeSet, has been added to Tree.java. When the TreeItem
has an image set in TreeItem.setImage(), the boolean is set to true.
This indicates that the pixbufRenderer is not to be resized any further
for that Tree.

Fixes for Bug 478915: this deals with an edge case where check box trees
created with TreeEditors were incorrectly placing their editor control.
This was apparent in the Run Configurations window, where the CCombo for
the "Auto-Start" and "Start Level" options was being placed at the
beginning of the tree, not in the correct columns. The fix necessary was
to take the position of the textRenderer into account when the size of
the pixbufRenderer hadn't been set -- causing it to have a position of
0.

Tested on Gtk3.16.7 and Gtk2.24. AllNonBrowserTests pass on Gtk3 and
Gtk2.

Change-Id: Ic9b39c44cfd0aff13f9591c511678061cda73a75
Signed-off-by: Eric Williams <ericwill@redhat.com>