Bug 480261: [GTK3][debug view] Tree icons missing in debug view, variables view
The fixes to bug 469277 include dynamically sizing Tree icons in Gtk3. 
This means that all trees have a pixbufRenderer size of 0x0 pixels until
an image is added.

Trees with the SWT.Virtual attribute had their fixed_height
property set to true, causing the size of the pixbufRenderer to remain
at 0 upon the initial creation of the tree. Removing this property
in Gtk3 allows the pixbufRenderer dimensions to be sized correctly and
thus fixes the bug present in the debug view.

Tested on Gtk3.16.6 and Gtk2.24. AllNonBrowser tests pass in Gtk3 and
Gtk2.

Change-Id: Idbf7fee787a4a74275b2deb68e2f0a458950e721
Signed-off-by: Eric Williams <ericwill@redhat.com>