Bug 480986: [Wayland] Workspace chooser won't launch
gtk_size_allocate() in Shell is called more often on Wayland. These
additional and unexpected calls trigger an infinite recursive call to
resizeBounds(), which causes non-resizable Shells/Dialogs to crash.

The fix is to only resize Shells that have the SWT.RESIZE attribute
enabled.

Tested on Gtk3.16 on Wayland, Gtk3.16 on X, and Gtk2.24. AllNonBrowser
JUnit test cases pass on Gtk3 running X. There are some irregularities
with test cases on Wayland, as outlined in bug 480893.

Change-Id: I55612d8486cdd512e275a8a7456e0197ed717c75
Signed-off-by: Eric Williams <ericwill@redhat.com>