Bug 481122: [GTK3.18+] some set/get Background/Foreground color methods have no effect
With changes in GTK3.18 and onward, gtk_style_context_get_color()
behaves differently. In order to correctly fetch the color we must first
save the GtkStyleContext, set its state, fetch the color, and then
restore the GtkStyleContext. Failure to do this on GTK3.18+ leads to
failures in the getForegroundColor() method.

The convenience method styleContextGetColor() in Display takes care of
the process.

Tested on GTK 3.18, 3.16, 3.14, 3.8, and 2.24. All foreground color
related tests on GTK3 now pass. GTK2 behaviour remains unchanged.

Change-Id: I0ca431a43c3eab30bd4f1a0d5282c3ae276d031c
Signed-off-by: Eric Williams <ericwill@redhat.com>