Bug 479998: [GTK3] Form/wizard Composites have incorrect background colors
As of GTK3.16 gtk_widget_override_background_color() is deprecated.
Unfortunately in GTK3.18 it has started to break the background colors
of SwtFixed containers. This is most notably visible in the workspace
selection and commit dialogs, where some background containers had the
default gray background color when they should have been white.

This fix overrides the draw signal in Composite, which allows for CSS
rendering of the background color using gtk_render_background(), which
is the recommended fix as per the GTK team's changes. This ensures that
every composite is drawn with the correct background color if it is set. 

The only additional side effect is that ToolBar background colors need
to be set with CSS also, since some SwtFixed containers that were
previously being overridden by ToolBar background colors now are
correctly displaying their own background. This means that ToolBars need
to be rendered with CSS as well. This can be accomplished very easily by
overriding the setBackgroundColor() method in ToolBar.java.

This change will only affect SWT using GTK3.16 and upward. Anything
below 3.16 still uses the old function (which works reliably on those
versions).

Change-Id: If9d1037fd24b9739e0c30ff794eb6fed0247d56f
Signed-off-by: Eric Williams <ericwill@redhat.com>