Bug 483112:[GTK3] TreeItems with multiple columns only show images in the first one
This patch fixes a regression from bug 469277, where Trees with multiple
columns only display images in the first column.

The problem lies with pixbufRenderers in other columns (other than the
first) not having their size set properly, resulting in no image being
displayed at all. The solution to this problem is to check columns as
images are being added to them, and compare the size of the
pixbufRenderer for that column to the size of the initial image set. If
there is a discrepancy, adjust the size of that column's pixbufRenderer
to the size of the initial image.

Tested on GTK3.18, GTK3.16, GTK3.14, GTK3.8, and GTK2.24. AllNonBrowser
JUnit tests pass on GTK2 and GTK3.

Change-Id: I560fecb7c0b9116510ced5ba7e8beddba9f16461
Signed-off-by: Eric Williams <ericwill@redhat.com>