Bug 483097: [GTK3.16+] gtk_widget_override_color is deprecated
This patch is an overhaul of foreground colors for GTK3.16+. Similar to
bug 483096, this bug aims to fix and prevent future bugs that will arise
from gtk_widget_override_color() being deprecated. This also fixes the
last failing foreground test case on GTK3.16.

The main feature of this patch is that it converts all (but one case) of
foreground colors to use GTK's CSS machinery. Except for
Control.checkForeground(), this means that SWT's background and
foreground colors are entirely GTK CSS compatible.

An important change in this patch affects both CSS backgrounds and
foregrounds: since Control uses the same GtkCssProvider for a widget, we
need to cache background and foreground values when using CSS. For
example, when setting the background color of a widget that already has
a foreground color, we need to load not only the "new" background color
but also re-load the foreground color to ensure it isn't lost. This is
accomplished via Display.gtk_create_css_color_string().

Tested on GTK3.18, 3.16, 3.14, and 2.24. No additional AllNonBrowser
JUnit tests fail on GTK3 and GTK2.

Change-Id: Ib80f7c51ba2c1c2c922a8f8b51c947f93ac01260
Signed-off-by: Eric Williams <ericwill@redhat.com>