Bug 483115: Scrollbar size wrongly reported
Previously, the macro GTK_SCROLLED_WINDOW_SCROLLBAR_SPACING()
responsible for returning scrollbar spacing was returning a value of 3.
This should be zero on the default GNOME theme.

According to the GTK team we should be accessing the "scrollbar-padding"
property differently. This style property belonging to GtkScrolledWindow
should now be accessed with a call to gtk_widget_get_style(). Using
this method instead will return the correct value. This eliminates some
extra padding that was being added to the bottom of the text editor in
Eclipse.

Tested on GTK3.18, GTK3.16, GTK3.14, and GTK2.24. AllNonBrowser JUnit
tests pass on GTK3 and GTK2.

Change-Id: I21c321688fa91af191e005f1aaf325a2cbae5cea
Signed-off-by: Eric Williams <ericwill@redhat.com>