Bug 464705: [Gtk3] Messed up contents when scrolling with GTK+ 3.16
In GTK3 we see scrollbar overlays enabled by default. This results in a
dotted line being drawn around the borders of a GtkScrolledWindow where
the scrollbars will be. These same dotted lines will sometimes be drawn
over the screen as you are scrolling.

This patch uses CSS to disable these overlays/dotted lines for all
GtkScrolledWindows. This results in somewhat cleaner scrolling and an
overall less cluttered appearance.

Tested on GTK3.18, 3.16, and 2.24. AllNonBrowser JUnit tests
pass on GTK3 and GTK2.

Change-Id: I0cddd5a188bd48e451042ac95543350f8b342bdf
Signed-off-by: Eric Williams <ericwill@redhat.com>