Bug 484696: [GTK] Empty repositories view links have white backgrounds
After bug 479998, an empty EGit repositories view still has mismatched
label colors. It turns out in this case the background color of the
parent composite was incorrect -- it should have been white (like on
GTK3.14 and earlier) instead of gray.

In this case the parent composite's background color is set to be that
of the TreeViewer which will display the future repository branches,
information, etc. The default color for Trees/Tables is white. The fixes
for bug 479998 and bug 479998 introduced a small regression which set
the default color of Table/Trees to be gray. The issue in this case does
not lie with the labels but with the color of the parent composite. With
Table/Trees returning their proper default background color (white, or
COLOR_LIST_BACKGROUND), the EGit label issue is fixed.

Tested on GTK3.18.6, 3.16, 3.14, and 2.24. AllNonBrowser JUnit tests
pass on GTK3 and GTK2.

Change-Id: I47fe95c1ad5fc87411c0f27ff344f6369c9aa0c6
Signed-off-by: Eric Williams <ericwill@redhat.com>