Bug 484729: [GTK3] Eclipse IDE consumes CPU when idle
Bug 479998 introduced a call to getClientArea() in Composite.gtk_draw().
This calls forceResize() in both the Composite and the parent Control,
which causes unnecessary overhead. It also causes a paint loop in
CTabFolder, with never ending calls to
gdk_cairo_region_create_from_surface(). The looped system calls to GDK
cause high CPU usage.

To remedy this we can use a GtkAllocation instead of calling
getClientArea(): this is far more efficient and does not trigger any
additional SWT machinery.

Tested on GTK3.18.6, 3.14, and 2.24. AllNonBrowser JUnit tests pass on
GTK3 and GTK2.

Change-Id: Ib1e5900e5c9339e9c0555a9436e8b2a9949b6372
Signed-off-by: Eric Williams <ericwill@redhat.com>