Bug 483096: [GTK3.16+] gtk_widget_override_background_color is deprecated
Bug 479998 introduced CSS background colors to Composite. This meant
setBackgroundColor() needed to be overridden in Composite to implement
this functionality. Since this patch implements CSS backgrounds for
all widgets in SWT, we can remove setBackgroundColor() from Composite.

A summary of fixes in this patch:
-widgets that are Composite only (i.e. SwtFixed containers) now have
 functioning background colors by overriding the gtk_draw() method
 (Link and Sash)
-Combo.setBackgrondColor() has been refactored/simplified
-some specific widgets with malfunctioning backgrounds have been fixed
 (Group, DateTime, ExpandBar, and Label)
-NullPointerException fixed in Table/Tree
-tweaking of background code in ToolBar
-fixed background image issues in ExpandBar and Label

This patch cleans up the setBackground() code and ensures background
color functionality for all widgets.

Tested on GTK3.18, 3.16, 3.14, and 2.24. AllNonBrowser JUnit tests pass
on GTK3 and GTK2.

Change-Id: I6f781541efbdbe5c27f6fbc5d441e5e24f782134
Signed-off-by: Eric Williams <ericwill@redhat.com>