Bug 485845 - Launching eclipse under wayland crashes.
GDK_WINDOWING_WAYLAND checks that wayland backend is compiled in but not
that we run on wayland. Also X11 will not be available on other backends
too thus assume GIO for everything not X11. GIO appeared in older
version than the one required by GTK3 (which added support for multiple
backends to be compiled in).

Change-Id: I0cd40ab1d257ea3c6c33b4d36d6c2ecfe05a1c40
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>