Bug 486069: [Wayland] NullPointerException in TabFolder when running ControlExample
This patch adds a missing logic check to prevent a GtkNotebook from
switching to a page that doesn't exist. Wayland seems to trigger an
early page switch, causing gtk_notebook_get_current_page() to return -1.
This will cause a NullPointerException in TabFolder.gtk_switch_page().

Tested on GTK3.19.6, and 3.18. AllNonBrowser JUnit tests pass on
GTK3.19.6 (with the exception of an unrelated previously failing test
case).

Change-Id: I720a7b08c0b49bebce15951f5d5deb00a2304839
Signed-off-by: Eric Williams <ericwill@redhat.com>