Bug 486334: [GTK3] Checkbox not visible in dialog
PART 2: This fixes behavior on GTK3.14.

Bug 478962 fixed some issues relating to GTK style context
machinery. These changes were in line with changes in GTK3.18, but the
older way of doing things needs to be preserved for versions of GTK
before 3.18. This will fix the bug on GTK3.14. The fix for GTK3.16 is
contained in part 1.

Tested on GTK3.18, 3.16, 3.14, 3.8, and 2.24. AllNonBrowser JUnit tests
pass on GTK2 and GTK3.

Change-Id: I4434d5ee27fb52fbf9a9b66c5a9fde4ebc32b692
Signed-off-by: Eric Williams <ericwill@redhat.com>