Bug 487712 - Snippet128 crashes on Fedora
Bug in ProgressBar due to calling gdk_window_process_updates which sends
expose events for redrawing purposes. But this event is implementation
detail in GTK3 and shouldn't be used from outside and playing with it
leads to race conditions with webkitgtk.
Also GtkProgressBar redraws properly now so trying to force redraw is
useless on GTK 3.

Change-Id: I22373b9e33c83759e0ff9dc2d0600fe39cacb91e
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>