Bug 487522: [GTK3.20] Entries/Text widgets have smaller heights
For GTK3.19 (and more importantly, 3.20) gtk_style_context_get_padding()
no longer returns correct values for GtkEntry widgets. GtkEntry widgets
in general no longer use padding and need to be sized using
width-for-height size management.

The solution here is to call computeNativeSize() on GTK3.19+ to
determine what the height of the entry should be. We can pass in the
width of the widget to computeNativeSize() which will give us the
preferred height based on the given width. This restores entries like
Quick Access, Search, etc. to their GTK3.18- sizes.

Tested on GTK3.19.6, 3.18, 3.16, 3.14, and 2.24. AllNonBrowser JUnit
tests pass on GTK3 and GTK2.

Change-Id: I076cadd66fa45d2dc14ceb78d624503999ca806f
Signed-off-by: Eric Williams <ericwill@redhat.com>