Bug 488226: [GTK3] Content Assist using Table/Tree has wrong selection background colors
On GTK3.14 and above, SWT.COLOR_LIST_SELECTION was returning gray
instead of the default (and correct) blue. This causes bugs in
Table/Tree, making highlighted TableItems/TreeItems difficult to read.
COLOR_LIST_BACKGROUND is unaffected.

System background colors like SWT.COLOR_* are fetched using
gtk_style_context_get_background_color() when SWT loads. This approach
is no longer valid as 1) it is deprecated on the GTK side and 2) it is
unreliable/not guaranteed to work. This patch only changes this behavior
for SWT.COLOR_LIST_SELECTION, but in the future other system colors will
need to be changed as well.

The fix in this case is to fetch the GTK theme colors of the currently
running system theme from GTK. This can be done using GtkCssProvider.
Once the currently running theme is fetched, the correct selection
background color can be extracted. If there is no selection background
color specified in the theme, then the fall-back is to load the color
via SWT's COLOR_LIST_SELECTION. This ensures that the correct color is
chosen according to GTK, should a theme specify one.

Tested on GTK3.8, 3.14, 3.16, and 3.18. AllNonBrowser JUnit tests pass
on GTK3 and GTK2.

Change-Id: Ic9aeb4d35efef961be1f724a2ad7dcc852c06453
Signed-off-by: Eric Williams <ericwill@redhat.com>