Bug 488497: [GTK3] Fix background irregularities in List
This patch converts List background color machinery to the same as Table
and Tree. List is a GtkTreeView on the GTK side, which means it needs to
be handled the same as Table and Tree when it comes to background
colors. This will fix some bugs with default List background and
selection background colors.

Tested on GTK3.18, 3.16, 3.14, and 2.24. AllNonBrowser JUnit tests pass
on GTK3 and GTK2.

Change-Id: I720761d4c104232296a7aff518f59fb872d975c5
Signed-off-by: Eric Williams <ericwill@redhat.com>