Bug 465056: [GTK3] Too narrow initial list size with scrollbars
This bug affects Tables with one column (i.e. no columns created).
Table.computeSize() will return an abnormally small size on GTK3. The
fix for this is to track and store the size of the widest TableItem in
Table.

This is accomplished by creating a maxWidth variable in Table. Anytime a
child TableItem has text or an image set, the width of that TableItem is
calculated, and compared to the value of maxWidth. If that TableItem is
wider than maxWidth, then maxWidth becomes the width of that TableItem.
When computeSize() is called on single column Tables, size.x is set to
maxWidth.

Tested on GTK3.18, 3.16, 3.14 and 2.24. AllNonBrowser JUnit tests pass
on GTK3 and GTK2.

Change-Id: Ib97c00e00c5b244a0e13901d976c3a0da444747f
Signed-off-by: Eric Williams <ericwill@redhat.com>