Bug 489379: [GTK3] Fix background irregularities in Text
This patch converts Text background color machinery to the same as Table
and Tree. Text is a GtkTextView on the GTK side, which means it needs to
be handled the same as Table and Tree when it comes to background
colors. This will fix some bugs with default Text background and
selection background colors. 

Tested on GTK3.18, 3.16, 3.14, and 2.24. AllNonBrowser JUnit tests pass
on GTK3 and GTK2.

Change-Id: Ic39714673c318ced4f70ab97c20694f9be28e667
Signed-off-by: Eric Williams <ericwill@redhat.com>