Bug 489953: [GTK3.14-] Button background tests fail
GTK3.14 and below does not have access to CSS parsing methods. There are
however a handful of widgets that use CSS background colors on GTK3.14.
Button is one of these, as CSS theming was introduced to Button to fix
some bugs.

This means that on GTK3.14 there is a test failure for
Button.getBackground() since SWT cannot parse GTK CSS on GTK3.14. The
solution to keep track of the background color and override
getContextBackground() in Button. This eliminates the last failing color
test case on GTK3.14

This modifies behavior for GTK3.14 and below. GTK3.16 behavior is
unaffected as CSS parsing machinery exists for 3.16 and up.

Tested on GTK3.18, 3.16, 3.14, and 2.24. AllNonBrowserTests pass on GTK3
and GTK2.