Bug 488498: [GTK3] Adapt remaining Display.initializeSystemColors() to use GTK theming
This patch converts the remaining SWT system color machinery to use GTK
CSS, similar to what bug 488226 did for SWT.COLOR_LIST_SELECTION.

The process is as follows: load the current GTK theme, parse it, seek
out the relevant values, parse and convert those values into GdkColor
objects for use in SWT. This procedure allows SWT to look more similar
to the native OS theme. It also allows for greater adaptability -- theme
values can be added over time to allow SWT to adapt to a wider variety
of GTK themes that might be popular/come packaged with different
distributions.

The old-style approach was to extract SWT system colors using
GtkStyleContext machinery, which is now deprecated. While it continues
to work for the time being, it will break eventually. The new method
which uses CSS is in line with newer versions of GTK.

Tested on GTK3.20, 3.18, 3.16, and 3.14. Tested using the Raleigh,
Adwaita, and Ambiance (Ubuntu) GTK themes. No additional AllNonBrowser
JUnit test failures occur on GTK3 and GTK2.

Change-Id: I73587469bac18a755d87b6e25a274194a59f758f
Signed-off-by: Eric Williams <ericwill@redhat.com>