Bug 490203: [GTK3] Unacceptable virtual Table performance
Bug 486334 introduced "dynamic" image sizing from Trees to Tables. This
was to fix the bug on GTK3.16 which caused by an earlier regression. In
these changes, a fix for bug 480261 was included. This was a Tree
specific bug which disabled fixed-height-mode for Trees to prevent
cropped icons.

Removing this for Tables was unnecessary as the bug hadn't been reported
against Table. This increases CPU usage as GTK cannot assume all rows
are the same height if fixed-height-mode is disabled. Re-enabling
fixed-height-mode for Tables with the SWT.VIRTUAL style set will restore
performance.

Tested on GTK3.14, 3.16, 3.18, and 2.24. AllNonBrowser JUnit tests pass
on GTK3 and GTK2.

Change-Id: Iacb3478d3edbda31fa896a66278d40fae1ef8551
Signed-off-by: Eric Williams <ericwill@redhat.com>