Bug 480261: [GTK3][debug view] Tree icons missing in debug view, variables view
This is a revised fix. The original fix was reverted due to performance
issues (see bug 490203).

Trees with SWT.VIRTUAL set as the style have fixed-height-mode enabled
on the GTK backend. This causes the GtkTreeView to have a fixed size.
Unfortunately this means that images/icons will be chopped off or
potentially non-existent. The fix is to re-create the renderers with the
appropriate sizes when an image is set for the first time. This is far
more efficient as it only runs once: after the initial image is set, its
size determines the size of all the other pixbuf renderers in the Tree.

This fix does not incur any additional performance hits like in bug
490203.

Tested on GTK3.18, 3.16, 3.14, and 2.24. No additional test failures for
AllNonBrowser JUnit tests on GTK2 and GTK3.

Change-Id: Ic105fd7198d16d429bc8f3385c5832f1c8e681e4
Signed-off-by: Eric Williams <ericwill@redhat.com>