Bug 490903 - Add ability to apply GTK CSS to widgets directly
More and more of Gtk functionality is exposed as CSS with no API for
getting/setting this information via function calls.
The implementation use Widget.setData with
org.eclipse.swt.internal.gtk.css as a key to mark it as both internal
and gtk specific, what it does is to apply the given css string via
gtk_css_provider_load_from_data to the widget context.

Change-Id: I8f120efa9f30fb8af6cad1e686591ad9e902ad2a
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>