Bug 480458: [GTK3.16+] Scroll bar hops around when scrolling
For GTK3.16, scroll bar styling changed to use overlays. This
means that the scroll bar is hidden unless the user hovers the mouse
inside the window.

Canvas has a scroll() method that handles the re-painting of its content
as the user scrolls. This includes fetching the visible region,
performing some calculations on it, and re-painting the
"updated" region (which includes portions of the previous region before
the scrolling happened).

During these calculations the scroll bar dimensions are included,
meaning the "old" positions of the scroll bar are re-painted as the user
scrolls, causing a flickering effect. The solution in this case is to
subtract the thickness of both scroll bars when performing the painting
calculations so that the scroll bar isn't redrawn constantly.

Tested on GTK3.20, 3.18, and 3.16. GTK3.14 and below (including GTK2) is
not affected. No additional AllNonBrowser JUnit test failures on GTK3
and GTK2.

Change-Id: Iba17be5ce919df78352662234a4bf0e02278e15a
Signed-off-by: Eric Williams <ericwill@redhat.com>