Bug 491324: [GTK3.16+] Shell with no decorations has black background
There is a call in Shell.createHandle() that sets a shell without
decorations to have a background color of black. While this seems to be
necessary on previous versions of GTK, it is unnecessary on 3.16+ as it
breaks the background color.

Tested on GTK3.20, 3.18, and 3.16. GTK3.14 or GTK2 is unaffected. No
additional AllNonBrowser JUnit test failures happen on GTK3 or GTK2.

Change-Id: Idec7f9027d65bbff7787642734f3f999d0f80871
Signed-off-by: Eric Williams <ericwill@redhat.com>