Bug 487467: [GTK3] org_eclipse_swt_widgets_Text.test_getTopIndex fails
This patch addresses a short-fall in GTK whereby Text widgets suffer
from a delay in calculating their top index. The root cause of the issue
lies in the fact that GtkTextView does line validation -- which runs in
an idle handler. This causes inaccurate values to be calculated. This
only affects GTK3.

The fix is to use a GtkTextMark on GTK3 to track the position of the
index in the buffer. GtkTextMarks are not subject to line validation.
This allows setTopIndex() and getTopIndex() to be called one after
another without issue. 

For cases when the user scrolls with the GUI (either by scrollbar or by
moving the cursor), getTopIndex() monitors the GtkAdjustment of the
vertical scrollbar. If there is any change in value, the old method of
calculation will be used. This method fetches the visible rectangle of
the GtkTextView and calculates the index of the top line.

Tested on GTK3.20, 3.18, 3.16, 3.14, and 2.24. This fixes a failing
AllNonBrowser test case on GTK3, and introduces no additional failures
on GTK2.

Change-Id: Ia015971786772579dfdca40d1091c6a40c6260cc
Signed-off-by: Eric Williams <ericwill@redhat.com>